Incremental update

+++ src/cli.rs
@@ -97,11 +97,11 @@ pub enum Commands {
         tabu_size_factor: f64,
 
         /// Speed type of drones.
-        #[arg(long, default_value_t = ConfigType::Low)]
+        #[arg(long, default_value_t = ConfigType::High)]
         speed_type: ConfigType,
 
         /// Range type of drones.
-        #[arg(long, default_value_t = ConfigType::Low)]
+        #[arg(long, default_value_t = ConfigType::High)]
         range_type: ConfigType,
 
         /// The number of trucks to override. Otherwise, use the default value.
@@ -125,8 +125,8 @@ pub enum Commands {
         fix_iteration: Option<usize>,
 
         /// The number of non-improved iterations before resetting the current solution = a2 * base
-        #[arg(long, default_value_t = 30)]
-        reset_after_factor: usize,
+        #[arg(long, default_value_t = 30.0)]
+        reset_after_factor: f64,
 
         /// The maximum size of the elite set = a3
         #[arg(long, default_value_t = 10)]
